requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
minio==7.2.0
PyPDF2==3.0.1
python-dotenv==1.0.0
sentence-transformers
chromadb
numpy
pandas
requests
tiktoken
transformers
torch
tqdm 